requirements.txt
lxml==5.1.0
pytest==8.2.0
psycopg2==2.9.9
requests==2.31.0
SQLAlchemy==2.0.29
fake-headers==1.0.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pyTelegramBotAPI==4.16.1
